feat: decouple image options view from dither page
