+++ data/ui/dither_page.blp
@@ -223,178 +223,6 @@ template $HalftoneDitherPage: Adw.BreakpointBin {
   }
 }
 
-Adw.Bin image_preferences_bin {
-  child: Adw.ToolbarView {
-    content: Adw.PreferencesPage {
-      Adw.PreferencesGroup options_group {
-        title: _("Options");
-
-        [header-suffix]
-        Gtk.MenuButton options_information_button {
-          valign: center;
-          popover: options_information_popover;
-          icon-name: "help-about-symbolic";
-          tooltip-text: _("Explanation");
-
-          styles [
-            "flat"
-          ]
-        }
-
-        Adw.ComboRow dither_algorithms_combo {
-          title: _("Dither Algorithm");
-          model: Gtk.StringList algorithms_stringlist {
-            strings ["Floyd-Steinberg", "Riemersma", "Bayer (ordered)"]
-          };
-        }
-
-        //Adw.ComboRow color_palette {
-        //  title: _("Color Palette");
-        //  model: Gtk.StringList color_palettes {};
-        //}
-
-        Adw.SpinRow color_amount_row {
-          title: _("Color Amount");
-          adjustment: Gtk.Adjustment color_amount_adjustment {
-            lower: 1;
-            upper: 256;
-            step-increment: 1;
-            page-increment: 5;
-          };
-          output => $on_color_amount_changed();
-        }
-
-        Adw.ComboRow export_format_combo {
-          title: _("Convert To");
-          model: Gtk.StringList image_formats_stringlist {};
-        }
-      }
-
-      Adw.PreferencesGroup adjustments_group {
-        title: _("Adjustments");
-
-        Adw.SpinRow brightness_row {
-          title: _("Brightness");
-          subtitle: _("Range -100 to 100");
-          adjustment: Gtk.Adjustment brightness_adjustment {
-            lower: -100;
-            upper: 100;
-            step-increment: 1;
-            page-increment: 5;
-          };
-          output => $on_brightness_changed();
-        }
-
-        Adw.SpinRow contrast_row {
-          title: _("Contrast");
-          subtitle: _("Range -100 to 100");
-          adjustment: Gtk.Adjustment contrast_adjustment {
-            lower: -100;
-            upper: 100;
-            step-increment: 1;
-            page-increment: 5;
-          };
-          output => $on_contrast_changed();
-        }
-      }
-
-      Adw.PreferencesGroup image_size_group {
-        title: _("Image Size");
-
-        Adw.SpinRow image_width_row {
-          title: _("Width");
-          subtitle: _("In pixels");
-          adjustment: Gtk.Adjustment image_width_adjustment {
-            lower: 1;
-            upper: 8192;
-            step-increment: 1;
-            page-increment: 5;
-          };
-          output => $on_image_width_changed();
-        }
-
-        // TODO: Remove option completely if there's no user feedback to bring it back
-        Adw.SwitchRow aspect_ratio_toggle {
-          visible: false;
-          title: _("Keep aspect ratio");
-
-          [suffix]
-          Gtk.MenuButton aspect_ratio_button {
-            valign: center;
-            popover: aspect_ratio_popover;
-            icon-name: "help-about-symbolic";
-            tooltip-text: _("Explanation");
-
-            styles [
-              "flat"
-            ]
-          }
-        }
-
-        // TODO: Remove option completely if there's no user feedback to bring it back
-        Adw.SpinRow image_height_row {
-          visible: false;
-          title: _("Height");
-          subtitle: _("In pixels");
-          adjustment: Gtk.Adjustment image_height_adjustment {
-            lower: 1;
-            upper: 8192;
-            step-increment: 1;
-            page-increment: 5;
-          };
-          output => $on_image_height_changed();
-        }
-      }
-    };
-
-    [bottom]
-    Gtk.Box {
-      halign: center;
-      Gtk.Button save_image_button {
-        label: _("_Save Image");
-        action-name: "app.save-image";
-        use-underline: true;
-
-        styles [
-          "pill",
-          "suggested-action"
-        ]
-      }
-
-      styles [
-        "sheet-toolbar"
-      ]
-    }
-  };
-}
-
-Gtk.Popover options_information_popover {
-  Gtk.Label {
-    margin-top: 6;
-    margin-bottom: 6;
-    margin-start: 6;
-    margin-end: 6;
-    wrap: true;
-    justify: center;
-    max-width-chars: 30;
-    use-markup: true;
-    label: _("To learn more about individual options, go to the <a href=\"https://github.com/tfuxu/Halftone/wiki/Options\">wiki page</a>.");
-  }
-}
-
-Gtk.Popover aspect_ratio_popover {
-  Gtk.Label {
-    margin-top: 6;
-    margin-bottom: 6;
-    margin-start: 6;
-    margin-end: 6;
-    wrap: true;
-    justify: center;
-    max-width-chars: 30;
-    label: _("This option allows you to preserve original aspect ratio when resizing an image.");
-  }
-}
-
 Gtk.FileDialog save_image_dialog {
   title: _("Save an Image in…");
   modal: true;
